S3 Data Source now fully working and Docling processing added for S3 files
- In sources/s3.py was rewritten to use s3fs directly and not se LlamaIndex S3Reader,
to download S3 files and then give to our DocumentProcessor (which processes with Docling) and then sets up LlamaIndex documents for LlamaLndex pipeline integration. Progress feedback callbacks are also setup in s3.py. The changes also fixed to give the feedback in terms of number of documents not count of chunks  as doc count.
- In backend.py handling of s3 progressing status was setup, and 2 un-needed processing status updating in general was removed
- ingest/manager.py was change to give better doc count progress feedback
- hybrid_system.py was changed to give specific feedback for youtube to be one document not chunk for doc count.
- In ProcessingTab.tsx for React, ProcessingTab.vue for Vue, and processing-tab.ts
for Angular, code was added to get the progress bar and other feed back to work for S3.
- In env-sample.txt configuration for S3 default parameters can be
- In requirements.txt include boto3 and s3fs for dealing with S3 instead of using LlamaIndex S3Reader